src/crawler/requirements.txt
nsepython==3.1.0
kafka-python==2.0.2
python-dotenv==1.0.0
pytz==2024.1